Refactor nested datastore classes and fix lifecycle issues in multitenancy
